Walker: make arithmetic expansion follow POSIX
POSIX requires signed long integer for arithmetic expansion. $? is
implemented in the interpreter class now because POSIX doesn't require
the type of $? to be long. It would cause conversion in many places if
we used long for $?.

+++ src/builtins/inherit_builtin.cpp
@@ -58,7 +58,7 @@ inline bool inherit_builtin::hasq(const std::string& value, const std::string& n
 // We do not support any QA warning
 int inherit_builtin::exec(const std::vector<std::string>& bash_args)
 {
-  _walker.set_value("ECLASS_DEPTH", _walker.resolve<int>("ECLASS_DEPTH") + 1);
+  _walker.set_value("ECLASS_DEPTH", _walker.resolve<long>("ECLASS_DEPTH") + 1);
 
   // find eclass directory
   std::string eclassdir;
@@ -114,8 +114,8 @@ int inherit_builtin::exec(const std::vector<std::string>& bash_args)
       _walker.set_value("INHERITED", _walker.resolve<std::string>("INHERITED") + " " + *iter);
   }
 
-  _walker.set_value("ECLASS_DEPTH", _walker.resolve<int>("ECLASS_DEPTH") - 1);
-  if(_walker.resolve<int>("ECLASS_DEPTH") > 0)
+  _walker.set_value("ECLASS_DEPTH", _walker.resolve<long>("ECLASS_DEPTH") - 1);
+  if(_walker.resolve<long>("ECLASS_DEPTH") > 0)
   {
     _walker.set_value("ECLASS", PECLASS);
   }
